% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rforecasting.R
\name{Rforecasting}
\alias{Rforecasting}
\title{Rforecasting}
\usage{
Rforecasting(df = .Last.value, Measure = "", Time = "",
  Pred.time = "", Last.period = "", launch.browser = False)
}
\arguments{
\item{df}{A dataframe input}

\item{Measure}{Specify value measures such as Revenue and Shipments}

\item{Time}{Specify time variables such as Year, Quarter, and Month}

\item{Pred.time}{Specify a number of years or quarters to predict}

\item{Last.period}{Specify the last time value. Year, Quarter}
}
\description{
Rforecasting
}
